+++ packages/cli/cli.ts
@@ -28,7 +28,17 @@ process.emit = function (name, data, ..._args) {
   return originalEmit.apply(process, arguments as unknown as Parameters<typeof process.emit>);
 };
 
-const packageJson = await import('./package.json', { with: { type: 'json' } });
+let packageJson: any;
+try {
+  packageJson = await import('./package.json', { assert: { type: 'json' } });
+} catch (err) {
+  if ((err as any).code === 'ERR_IMPORT_ATTRIBUTE_MISSING') {
+    packageJson = await import('./package.json', { with: { type: 'json' } });
+  } else {
+    console.error('💥 Error occured:', '\n\n', err);
+    process.exit(1);
+  }
+}
 
 const { blue, bold, green, italic, magenta, red } = colors;
 const { ModuleKind, ScriptTarget, transpileModule } = ts;
@@ -192,15 +202,34 @@ program
   .description('Update all `@remix-pwa/*` packages to latest')
   .option('-p, --packages <packages...>', 'Individual packages to update')
   .option('-r, --root <root>', "Location of app's root directory (where package.json is located)", '.')
+  .option('-D, --dev', "Update packages to the latest 'dev' version")
   .action(async options => {
     const root = resolve(process.cwd(), options.root);
     const packagesToUpdate = options.packages ?? [];
+    const TAG = options.dev ? 'dev' : 'latest';
 
-    const packageJSON = (
-      await import('file://' + fileURLToPath(`file:///${resolve(root, 'package.json').replace(/\\/g, '/')}`), {
-        with: { type: 'json' },
-      })
-    ).default;
+    let packageJSON: any;
+    console.log(blue('🔃 Reading package.json file...'));
+
+    try {
+      packageJSON = (
+        await import('file://' + fileURLToPath(`file:///${resolve(root, 'package.json').replace(/\\/g, '/')}`), {
+          assert: { type: 'json' },
+        })
+      ).default;
+    } catch (err) {
+      if ((err as any).code === 'ERR_IMPORT_ATTRIBUTE_MISSING') {
+        packageJSON = (
+          await import('file://' + fileURLToPath(`file:///${resolve(root, 'package.json').replace(/\\/g, '/')}`), {
+            with: { type: 'json' },
+          })
+        ).default;
+      } else {
+        console.error(red(bold('💥 Error occured whilst reading package.json file:')), '\n\n', err);
+        process.exit(1);
+      }
+    }
+
     const dependencies = packageJSON.dependencies;
     const devDependencies = packageJSON.devDependencies;
 
@@ -217,7 +246,7 @@ program
 
       console.log(blue('Found the following `@remix-pwa/*` packages:'));
       allPackages.forEach(pkg => pkg !== '|' && console.log(green(`- ${pkg}`)));
-      console.log('\n', blue(`🚀 Updating all packages to ${italic('latest')}...`));
+      console.log('\n', blue(`🚀 Updating all packages to ${italic(TAG)}...`));
 
       allPackages.forEach(pkg => {
         if (pkg === '|') return;
@@ -226,13 +255,13 @@ program
 
         if (depType === 'dep') {
           try {
-            execSync(`npm i ${pkg}@latest`);
+            execSync(`npm i ${pkg}@${TAG}`);
           } catch (err) {
             console.error(`${red(`💥 Error occured whilst installing ${pkg}:`)}\n\n${err}`);
           }
         } else {
           try {
-            execSync(`npm i -D ${pkg}@latest`);
+            execSync(`npm i -D ${pkg}@${TAG}`);
           } catch (err) {
             console.error(`${red(`💥 Error occured whilst installing ${pkg}:`)}\n\n${err}`);
           }
@@ -247,7 +276,7 @@ program
 
     console.log(blue('Confirmed and updating the following packages:'));
     packagesToUpdate.forEach(pkg => console.log(green(`- @remix-pwa/${pkg}`)));
-    console.log('\n', blue(`🚀 Updating all confirmed packages to ${italic('latest')}...`));
+    console.log('\n', blue(`🚀 Updating all confirmed packages to ${italic(TAG)}...`));
 
     packagesToUpdate.forEach(dep => {
       let depType: 'dep' | 'devDep' = 'dep';
@@ -257,13 +286,13 @@ program
 
       if (depType === 'dep') {
         try {
-          execSync(`npm i ${dep}@latest`);
+          execSync(`npm i ${dep}@${TAG}`);
         } catch (err) {
           console.error(`${red(`💥 Error occured whilst installing ${dep}:`)}\n\n${err}`);
         }
       } else {
         try {
-          execSync(`npm i -D ${dep}@latest`);
+          execSync(`npm i -D ${dep}@${TAG}`);
         } catch (err) {
           console.error(`${red(`💥 Error occured whilst installing ${dep}:`)}\n\n${err}`);
         }