+++ packages/cli/CHANGELOG.md
@@ -1,3 +1,32 @@
+## @remix-pwa/cli 1.2.5 (2024-08-11)
+
+
+### Bug Fixes
+
+* **cli:** added new `--dev` flag to the `upgrade` command - for updating to latest `dev` candidate d4f4414
+
+## @remix-pwa/cli 1.2.5-dev.1 (2024-08-11)
+
+
+### Bug Fixes
+
+* **cli:** added new `--dev` flag to the `upgrade` command - for updating to latest `dev` candidate d4f4414
+
+## @remix-pwa/cli 1.2.4 (2024-07-21)
+
+
+### Bug Fixes
+
+* **cli:** added cross-support for Node pre-21 & Node 21.0.7+ b5f1c1b
+* **cli:** upgraded dependencies 0a7470a
+
+## @remix-pwa/cli 1.2.4-dev.2 (2024-07-21)
+
+
+### Bug Fixes
+
+* **cli:** added cross-support for Node pre-21 & Node 21.0.7+ b5f1c1b
+
 ## @remix-pwa/cli 1.2.4-dev.1 (2024-07-20)