+++ packages/sw/CHANGELOG.md
@@ -1,3 +1,24 @@
+## @remix-pwa/sw 3.0.9 (2024-08-11)
+
+
+### Bug Fixes
+
+* **sw:** added `matchOptions` for adding querying options to cache operations c87e507
+
+## @remix-pwa/sw 3.0.9-dev.1 (2024-08-11)
+
+
+### Bug Fixes
+
+* **sw:** added `matchOptions` for adding querying options to cache operations c87e507
+
+## @remix-pwa/sw 3.0.8 (2024-07-21)
+
+
+### Bug Fixes
+
+* **sw:** upgraded dependencies 197fc70
+
 ## @remix-pwa/sw 3.0.8-dev.1 (2024-07-20)